Edit OOP and nested methods
