+++ src/test/java/TaskManagerTest.java
@@ -4,8 +4,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class TaskManagerTest {
 
     @Test
-    public void taskManagerTest() {
-        TaskManager taskManager = new TaskManager();
-        assertEquals("Bye. Hope to see you again soon!", taskManager.executeCommand("bye"));
+    public void taskListTest() {
+        TaskList taskList = new TaskList();
+        Storage storage = new Storage();
+        Parser parser = new Parser();
+
+        assertEquals("Bye. Hope to see you again soon!", parser.parse("bye", taskList, storage));
     }
 }
